feat: hide list btns

+++ packages/form-render/src/core/RenderChildren/RenderList/TableList.js
@@ -52,31 +52,33 @@ const TableList = ({
     };
   });
 
-  columns.push({
-    title: '操作',
-    key: '$action',
-    fixed: 'right',
-    width: 60,
-    render: (value, record, idx) => {
-      return (
-        <Popconfirm
-          title="确定删除?"
-          onConfirm={() => deleteItem(idx)}
-          okText="确定"
-          cancelText="取消"
-        >
-          <a>删除</a>
-        </Popconfirm>
-      );
-    },
-  });
+  if (!props.hideDelete) {
+    columns.push({
+      title: '操作',
+      key: '$action',
+      fixed: 'right',
+      width: 60,
+      render: (value, record, idx) => {
+        return (
+          <Popconfirm
+            title="确定删除?"
+            onConfirm={() => deleteItem(idx)}
+            okText="确定"
+            cancelText="取消"
+          >
+            <a>删除</a>
+          </Popconfirm>
+        );
+      },
+    });
+  }
 
   return (
     <>
       <div className="w-100 mb2 tr">
-        <Button type="primary" size="small" onClick={addItem}>
+        {!props.hideAdd && <Button type="primary" size="small" onClick={addItem}>
           新增
-        </Button>
+        </Button>}
         {Array.isArray(props.buttons)
           ? props.buttons.map((item, idx) => {
               const { callback, text, html } = item;
